Add new theme
I look it to google for bootstrap themes, and reach this page:

http://bootswatch.com/

I take the first oen and replace it, seems like all is working except
for the fonts.

Hope you like it.

Look at that place and play arround with the different themes.

Look at the changes in this commit, to see what I did.